Search bar functioning; Fixed error state reset bug in ChartArea, and some render logic improvements; State lift-up into FinancialDashboard, as single source of truth to feed into SearchBar and ChartArea
