+++ frontend/src/components/Navbar/Navbar.jsx
@@ -11,7 +11,7 @@ import {
 import { Link } from "react-router-dom";
 import { useEffect, useState } from "react";
 
-function Navbar({setTicker}) {
+function Navbar(props) {
   return (
     <Box sx={{ display: "flex", justifyContent: "center" }}>
       <Box
@@ -30,7 +30,7 @@ function Navbar({setTicker}) {
           height="100"
         />
         <img src={alphaLogo} alt="AlphaView Logo" width="300" height="40" />
-        <SearchBar setTicker={setTicker} sxProps={{marginLeft: 2}}/>
+        {props.children}
       </Box>
       <Box sx={{ display: "flex", alignItems: "center", marginRight: 5 }}>
         <Button
@@ -48,71 +48,5 @@ function Navbar({setTicker}) {
     </Box>
   );
 }
-
-function SearchBar({setTicker, sxProps = {}}) {
-  
-  const [ symbols, setSymbols ] = useState([]);
-  const [ isError, setError ] = useState(false);
-  const [ isLoading, setLoading ] = useState(false);
-  const [ selectedSymbol, setSelectedSymbol ] = useState(null);
-
-  const submitHandler = (event) => {
-      event.preventDefault();
-      setTicker(selectedSymbol);
-      console.log(`Ticker submitted: ${selectedSymbol}`)
-  };
-
-  useEffect(() => {
-    const fetchSymbols = async () => {
-      try {
-        setLoading(true);
-        const res = await fetch(
-          `${import.meta.env.API_URL}/symbol/list/latest`,
-        );
-
-        if (!res.ok)
-          throw new Error(`HTTP ${res.status}: Failed to fetch symbol list`);
-
-        const symbolJSON = await res.json();
-        const symbolList = [];
-        for (const rowObj of symbolJSON) {
-          symbolList.push(rowObj.symbol);
-        }
-        setSymbols(symbolList);
-      } catch (error) {
-        console.log(error);
-        setError(true);
-      }
-    };
-    fetchSymbols();
-    setLoading(false);
-  }, []);
-
-  return (
-    <>
-      {isLoading && <p>Loading search...</p>}
-      {isError && <p>Error loading ticker search.</p>}
-      {!isError && !isLoading && (
-      <Box component="form" onSubmit={submitHandler} noValidate>
-      <Autocomplete
-        options={symbols}
-        value={selectedSymbol}
-        onChange={(event, newValue) => setSelectedSymbol(newValue)}
-        renderInput={(params) => (
-          <TextField
-            {...params}
-            label="Ticker"
-            sx={{
-              "& .MuiOutlinedInput-root": { borderRadius: 5 },
-              width: 130,
-              ...sxProps
-            }}
-          />
-        )}
-      />
-      </Box>
-      )}
-  </>
-);}
 
 export default Navbar;
