+++ frontend/src/routes/FinanceDashboard/FinanceDashboard.jsx
@@ -1,4 +1,5 @@
 import Navbar from "@components/Navbar/Navbar.jsx";
+import SearchBar from "@components/SearchBar/SearchBar.jsx";
 import NewsFeed from "./NewsFeed.jsx";
 import ChartArea from "./ChartArea.jsx";
 import MetricsCard from "./BottomBarMetrics.jsx";
@@ -7,24 +8,32 @@ import styles from "./FinanceDashboard.module.css";
 import { useState } from "react";
 
 function FinanceDashboard() {
-    
-    const [ stockSymbol, setStockSymbol ] = useState(null);
-    
-    const handleTickerInput = (symbol) => {
-       setStockSymbol(symbol);
-       console.log(`Stock symbol changed at app level: ${stockSymbol}`)
-    };
+  const [stockSymbol, setStockSymbol] = useState(null);
 
-    return (
-        <div className={styles["pg"]}>
-            <Navbar setTicker={handleTickerInput}/>
-            <div className={styles["pg-content"]}>
-                <NewsFeed length={30} cardStyles={{ width: "35%", height: "100%" }} />
-                <ChartArea symbol={stockSymbol} cardStyles={{ width: "100%", height: "100%" }} />
-                <WatchlistCard cardStyles={{ width: "35%", height: "100%" }} />
-            </div>
-        </div>
-    );
+  function handleTickerInput(symbol) {
+    setStockSymbol(symbol);
+    console.log(`Stock symbol changed at app level: ${symbol}`);
+  }
+
+  return (
+    <div className={styles["pg"]}>
+      <Navbar>
+        <SearchBar
+          value={stockSymbol}
+          onTickerSelect={handleTickerInput}
+          sxProps={{ marginLeft: 2 }}
+        />
+      </Navbar>
+      <div className={styles["pg-content"]}>
+        <NewsFeed length={30} cardStyles={{ width: "35%", height: "100%" }} />
+        <ChartArea
+          symbol={stockSymbol}
+          cardStyles={{ width: "100%", height: "100%" }}
+        />
+        <WatchlistCard cardStyles={{ width: "35%", height: "100%" }} />
+      </div>
+    </div>
+  );
 }
 
 export default FinanceDashboard;
